Better recent access only ranking.
Calculate recent access only rank from last access timestamp, so that
the higher the rank, the more recent the file has been accessed.
Previous way of doing recent access only ranking was broken with
Bayesian Ranking. This way of doing it fixes it. Also result sorting
could be simplified with this modification.