Tweak `--sanitize' code.
This should eliminate most if not all cases of side-effects when
`eval'ing.

The new sed code does two things:

* Remove `$(.*)' and "`.*`" style command substitutions.
* Remove all `[|&;<>$`]' characters that are not escaped by `\'.

Caveat: quoted characters `[|&;<>$`]' are still removed, causing
filenames with `[|&;<>$`]' not to be tracked if they are quoted. But
there is no easy way to solve this, since detecting quoting state is
pretty hard to do in plain sed.