New matching method via grep.
The new method offers three level of matching: default,
case-insensitive, and fuzzy.

* Default matching mode:

1. Queries must appear *in order* in the path.
2. The last query must match the last segment of the path.

* Case-insensitive mode:

1. Inherent 1, 2 from the default matching mode
2. Case-insensitive

* Fuzzy mode:

1. Inherent 1, 2 from case-insensitive mode
2. Extra-characters allowed between query characters

This idea is borrowed from autojump.

Performance note:

Due to my testing on various platforms, performance is on par with the
previous implementation. It performs better when using older grep
implementations that does not take locale into account. It performs
better when hard disk access is not fast, for instance when there are
removable drives. However, it performs slower on some query strings when
all three matching modes have to be tried.