Escape BRE special characters, translate `$'.
`.' is often used as a delimiter in filenames, it is un-intuitive for it
to be a wild card. With `.' being literal, `*' have no much use to be
special since people don't know it's BRE they are writing, and it's also
useless for `*' to have its shell globing meaning, since all the user
have to type is ` ' to represent `*'. It is also counter-productive to
use `[' grouping in fasd queries. Therefore, escape all BRE special
characters (`.[\*'). `^' does not have to be escaped since users' query
does not get to be the first character. `$' could be useful having a
special meaning if it is the last character. Therefore, preserve its
meaning by translate it into its equivalent (`|') in our context.