Provide a solution for issue #95
The problem is that the BSD version of `grep` has the peculiar behavior of
printing the string `(standard input)` on matches that occur on standard
input. This results in that literal string being included with every
match. Which disturbs the matching logic of `fasd`.

For example, the code below is affected:

```
[ -${typ:-e} "${line%%\|*}" ] && printf %s\\n "$line"
```

On a match, the `${line%%\|*}` line will contain `(standard
input):some_filename`. As a result, the file test operator (`-e`, `-d`,
etc) will not succeed and so the query will not return anything :(

This fix makes the name of the `grep` utility a configurable parameter.
This would allow somebody to install the GNU version of grep and define
`_FASD_GREP=PATH_TO_GNU_GREP` inside `.fasdrc`.

Problem with running on a mac #95