Amended fasd_cd to output pwd following z, zz, etc.
The pwd command is conditional on "$_FASD_PWD" == "1"; default action
if it is not set or set to anything else is not to output pwd.

I got this from the 'z' project (which I otherwise did not get on
with). It is useful for people with a $PS1 that does not show the
/entire/very/long/folder/path, like me with a not-so-wide terminal.

I also edited the README.md to include info on the env var.

This is my first PR, so if it’s out of order or not acceptable, please
let me know and I shan’t do it again!